feat: add minting page for STRKP tokens
* A new component `MintPage` was created that allows users to convert STRK tokens to STRKP.
* The `TokenMint` component was implemented to handle the minting logic and notifications.
* The token exchange page, `SwapPage`, was removed as it is no longer needed.
* Informative sections about STRKP tokens and minting guidelines were added.
